Allow all combinations of Inkbunny ratings
Since you might have a submission that is both sexual and violent or
something. To fit with the one-dimensional idea of ratings, this now
offers all combinations as separate buttons. It could be done as
checkboxes instead too, but this way works fine without reinventing the
whole ratings concept in multiple dimensions.