Don't pre-select blank default template part
If the default part of a template is completely empty, the user probably
doesn't want to use it to clobber their inputs with nothingness. So now
a blank default part is deselected by default when applying a template
to a submission, which seems like the more intuitive behavior. If the
user really wants the clobbering behavior, they can manually check it.

An even cooler implementation would probably let a user pick which
pieces of a template are supposed to apply to avoid this kind of coarse
guesswork, but this works okay without totally reimagining templates.