Allow posting to Discord forum channels
Discord forum channels require a `thread_name` to be passed as a query
parameter of the first post, with subsequent posts in the same thread
using a `thread_id` parameter in the body. Annoyingly, they validate it
both ways, so we can't just always send those parameters, the user has
to tick a box in the account settings if the target is a forum channel
instead of a regular message channel.

This also changes file posting to stick everything into a single post
instead of splitting it up into a notification message and one message
per file. Presumably Discord didn't use to support that, but it does
now, using the `files[n]` instead of `file` to attach files and a
`payload_json` parameter to attach nested data into a multipart form.