Replies for Mastodon and Bluesky (#249)
* Allow specifying reply to URL for Mastodon posts

Because I wanna be able to post threads. This interface isn't too
comfortable, since it requires manual copying of the parent post URL.
But letting your wire up a parent post from within PostyBirb is left as
an exercise for someone else or future me, this is currently already
useful as a minimum viable product.

* Allow specifying reply to URL for Bluesky posts

Analogous to the Mastodon implementation in the previous commit, except
more complicated because Bluesky.