Properly validate Bluesky description length
It wasn't using the plain text version of the description, so its
character count was inflated by counting the raw HTML. This is rectified
now, by using the same kind of code that other services use.

The error message also now says "characters" again instead of
"graphemes", since graphemes are what humans interpret as characters,
even if the computer may see them as a collection of characters.