Make content warnings part of the default section
Because I'm getting tired of repeating it for a bunch of sites.

The content warning will automatically be applied to Itaku, Mastodon,
Misskey, Pixelfed and Pleroma. They can also overwrite the default via
toggle switch, similar to how it works for descriptions. Probably mostly
necessary for Itaku with its dinky 30 character limit.

The content warning can be included into the description using the {cw}
shortcut. This is mostly useful for sites like Bluesky and Telegram,
which support spoilers, but don't have a dedicated content warning
field.

Backward compatibility with existing posts should be preserved: if the
overwrite option is undefined, it will be guessed by the presence of a
non-empty content warning text.

This also now consistently presents the field as "content warning" to
the user, rather than revealing the internal "spoiler text" name. I've
seen confused users on Mastodon thinking that PostyBirb didn't support
content warnings because of the misnomer.