Allow posting multiple files to Itaku
By uploading them as separate images and then creating a post that
references them. This is how multi-image posting works on Itaku, it's
really strange.

If something goes wrong with posting the images, the user will end up
with duplicates. Not much we can really do about it though and that's
always a risk with retrying a post anyway, since the user might get a
failed response even though the image posted fine.

This requires "share on feed" to be turned on, since posts are always
shared on the feed.