Fix description length validation and tag limiting
The validation functions would get passed the raw description as entered
by the user and would make some half-hearted effort to parse them into
something sensible, but this completely breaks down in the face of
shortcuts like {default}, username shortcuts that get turned into long
links etc. Now those functions get the actually parsed description
passed to them, the only unresolved shortcut in them being the `{tags}`
one, since the length of that is variable and up to the validation
function to deal with and tell the user if tags didn't fit.

Since #286, tags insertion is also pretty much busted entirely, since it
runs at the wrong time, so it may insert too many tags and then cause an
error down the line because the resulting description is too long. That
insertion logic has now been moved to the end of the description
processing, where it belongs.