Add Bluesky video support
The implementation is conceptually simple: it checks if we're allowed to
upload video, uploads it, then polls the processing status until
completion and then posts the post. What makes this difficult is that
the atproto library is completely busted in three dimensions.

First, the main module file contains what looks to me like a syntax
error. The TypeScript compiler agrees, so you can't even compile it. The
workaround is to avoid this main file and instead import stuff piecemeal
instead.

Then, while they allow you to specify a fetch handler, they still
unconditionally access a bunch of browser-specific stuff (Request,
Response, FormData) in the global scope, requiring us to jam those in
there so that it's happy.

Finally, the implementation of the video endpoints in the API is plain
wrong, missing required arguments (that you can't specify anyway because
then validation fails), not requesting the required access tokens and
hitting the wrong endpoint altogether. So we instead punt to manual
fetching in the same way that the website does.