Do Bluesky login check in electron-app only
After rebasing to the main repository's develop branch, Bluesky login in
the UI is suddenly failing with some kind of garbage "url must be a
string not an object" error when we aren't passing any such parameter in
the first place. I don't know what that's about, but doing it in
electron-app works fine, so I changed the UI code to defer to that
instead. This also improves the error messages so that the user isn't
left with the page seemingly not reacting to their input at all.