Warn if Bluesky password isn't app-password-esque
App passwords have a pretty regular format, so we can check for that and
warn the user that they enter something else, probably their account
password. Since the app password format could conceivably change in the
future, we leave it at a warning and don't refuse saving in that case.