Fix Bsky login check (#382)
* Handle Bluesky login check without credentials

It accessed undefined data and fell on its face, now it returns
successfully with loggedIn set to false.

* Do Bluesky login check in electron-app only

After rebasing to the main repository's develop branch, Bluesky login in
the UI is suddenly failing with some kind of garbage "url must be a
string not an object" error when we aren't passing any such parameter in
the first place. I don't know what that's about, but doing it in
electron-app works fine, so I changed the UI code to defer to that
instead. This also improves the error messages so that the user isn't
left with the page seemingly not reacting to their input at all.

* Warn if Bluesky password isn't app-password-esque

App passwords have a pretty regular format, so we can check for that and
warn the user that they enter something else, probably their account
password. Since the app password format could conceivably change in the
future, we leave it at a warning and don't refuse saving in that case.